requirements.txt
aioconsole==0.3.2
aiodns==3.0.0
cffi==1.14.6
cycler==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.2
numpy==1.21.2
Pillow==8.3.1
pkg_resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.0.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
six==1.16.0
slixmpp==1.7.1
